+++ libs/sid/libsidplayfp/src/builders/residfp-builder/residfp/resample/SincResampler.h
@@ -1,7 +1,7 @@
 /*
  * This file is part of libsidplayfp, a SID player engine.
  *
- * Copyright 2011-2024 Leandro Nini <drfiemost@users.sourceforge.net>
+ * Copyright 2011-2013 Leandro Nini <drfiemost@users.sourceforge.net>
  * Copyright 2007-2010 Antti Lankila
  * Copyright 2004 Dag Lem <resid@nimrod.no>
  *
@@ -74,7 +74,7 @@ private:
 
     int outputValue;
 
-    int sample[RINGSIZE * 2];
+    short sample[RINGSIZE * 2];
 
 private:
     int fir(int subcycle);
